Move execution classes into an Execution namespace and folder
