+++ Source/Machine.VSTestAdapter.Specs/TestAdapter/When_discovering_an_assembly_that_results_in_an_exception.cs
@@ -5,6 +5,7 @@ using Microsoft.VisualStudio.TestPlatform.ObjectModel.Logging;
 using System;
 using System.Collections.Generic;
 using Machine.VSTestAdapter.Discovery;
+using Machine.VSTestAdapter.Execution;
 
 namespace Machine.VSTestAdapter.Specs.TestAdapter
 {
