+++ Source/Machine.VSTestAdapter/MSpecTestAdapter.cs
@@ -7,7 +7,6 @@ using System.Diagnostics;
 using System.Linq;
 using System.Reflection;
 using Machine.VSTestAdapter.Discovery;
-using Machine.VSTestAdapter.Discovery.BuiltIn;
 
 namespace Machine.VSTestAdapter
 {
