+++ Source/Machine.VSTestAdapter/Machine.VSTestAdapter.csproj
@@ -76,18 +76,18 @@
     <Compile Include="Discovery\Cecil\IDelegateFieldScanner.cs" />
     <Compile Include="Discovery\Cecil\ItDelegateFieldScanner.cs" />
     <Compile Include="MspecTestAdapterExecutor.cs" />
-    <Compile Include="MSpecVSRunnerManager.cs" />
-    <Compile Include="ISpecificationExecutor.cs" />
+    <Compile Include="Execution\MSpecVSRunnerManager.cs" />
+    <Compile Include="Execution\ISpecificationExecutor.cs" />
     <Compile Include="Discovery\ISpecificationDiscoverer.cs" />
     <Compile Include="MSpecTestAdapter.cs" />
     <Compile Include="MSpecTestAdapterFactory.cs" />
     <Compile Include="Discovery\MSpecTestCase.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="RunStats.cs" />
+    <Compile Include="Execution\RunStats.cs" />
     <Compile Include="Discovery\Cecil\ScopedAssemblyResolver.cs" />
     <Compile Include="Discovery\Cecil\CecilSpecificationDiscoverer.cs" />
-    <Compile Include="SpecificationExecutor.cs" />
-    <Compile Include="SpecificationRunListener.cs" />
+    <Compile Include="Execution\SpecificationExecutor.cs" />
+    <Compile Include="Execution\SpecificationRunListener.cs" />
     <Compile Include="SpecTestHelper.cs" />
     <Compile Include="Strings.Designer.cs">
       <AutoGen>True</AutoGen>
