+++ Source/Machine.VSTestAdapter/MspecTestAdapterExecutor.cs
@@ -5,6 +5,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Diagnostics;
+using Machine.VSTestAdapter.Execution;
 
 namespace Machine.VSTestAdapter
 {
